+++ core/playback/mpv/track.go
@@ -8,7 +8,6 @@ package mpv
 import (
 	"fmt"
 	"os"
-	"runtime"
 	"time"
 
 	"github.com/dexterlb/mpvipc"
@@ -101,40 +100,6 @@ func (t *MpvTrack) Pause() {
 	err := t.Conn.Set("pause", true)
 	if err != nil {
 		log.Error("Error pausing track", "track", t, err)
-	}
-}
-
-func (t *MpvTrack) Close() {
-	// Windows automatically handles closing
-	// and cleaning up named pipe
-	if runtime.GOOS == "windows" {
-		return
-	}
-	log.Debug("Closing resources", "track", t)
-	t.CloseCalled = true
-	// trying to shutdown mpv process using socket
-	if t.isSocketFilePresent() {
-		log.Debug("sending shutdown command")
-		_, err := t.Conn.Call("quit")
-		if err != nil {
-			log.Error("Error sending quit command to mpv-ipc socket", err)
-
-			if t.Exe != nil {
-				log.Debug("cancelling executor")
-				err = t.Exe.Cancel()
-				if err != nil {
-					log.Error("Error canceling executor", err)
-				}
-			}
-		}
-	}
-
-	if t.isSocketFilePresent() {
-		log.Debug("Removing socketfile", "socketfile", t.IPCSocketName)
-		err := os.Remove(t.IPCSocketName)
-		if err != nil {
-			log.Error("Error cleaning up socketfile", "socketfile", t.IPCSocketName, err)
-		}
 	}
 }
