+++ core/playback/mpv/track_close_win.go
@@ -1,0 +1,8 @@
+//go:build windows
+
+package mpv
+
+func (t *MpvTrack) Close() {
+	// Windows automatically handles closing
+	// and cleaning up named pipe
+}
